Use numdifftools for derivatives, add function CI
Providing the gradient and Hessian is not a requirement anymore.
If not provided, they will be computed using numdifftools.

A new function is added to compute confidence intervals of functions
of parameters. The general method is also described in the paper.
This is mostly a convencience function, but smart computation of the
Hessian makes things faster and easier.